Change .ts to .js
